# public/openapi.yml
openapi: 3.0.0
info:
  title: AnkerSolixE1600-APIServices
  version: '1.0.5'
  description: |
    API Services for the Anker Solix Solarbank E1600 for integration with Smart-Home or Energy-Management-Systems.

    GitHub: [AnkerSolixE1600-APIServices](https://github.com/energychain/AnkerSolixE1600-APIServices)

    Development funded by buyers of [Corrently Energy](https://corrently.energy/).
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

servers:
      - url: http://192.168.88.3:3001
        description: Local runtime/server

paths:
  /setEnergy/{site_id}:
    post:
      summary: Sets the energy for a specific site.
      parameters:
        - in: path
          name: site_id
          required: true
          description: The ID of the site.
          schema:
            type: integer
        - in: query
          name: site_id
          schema:
            type: integer
          description: The ID of the site.
        - in: query
          name: power
          schema:
            type: integer
          description: The power value.
        - in: body
          name: body
          required: false
          schema:
            type: object
            properties:
              power:
                type: integer
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error
  /schedule/{site_id}:
    get:
      summary: Retrieves the schedule for a specific site.
      parameters:
        - in: path
          name: site_id
          required: true
          description: The ID of the site.
          schema:
            type: integer
        - in: query
          name: site_id
          schema:
            type: integer
          description: The ID of the site.
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error
    post:
      summary: Sets the schedule for a specific site.
      parameters:
        - in: path
          name: site_id
          required: true
          description: The ID of the site.
          schema:
            type: integer
        - in: query
          name: site_id
          schema:
            type: integer
          description: The ID of the site.
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              ranges:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    start_time:
                      type: string
                    end_time:
                      type: string
                    turn_on:
                      type: boolean
                    appliance_loads:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          power:
                            type: integer
                          number:
                            type: integer
              min_load:
                type: integer
              max_load:
                type: integer
              step:
                type: integer
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error
  /getHomepage:
    get:
      summary: Retrieves the site homepage.
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error
  /restart:
    all:
      summary: Restarts the service.
      responses:
        '503':
          description: Service Restarting
  /login:
    post:
      summary: Authenticates the user with Anker Cloud Credentials and retrieves the site homepage.
      requestBody:
        description: User credentials
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ANKER_USERNAME:
                  type: string
                ANKER_PASSWORD:
                  type: string
                ANKER_COUNTRY:
                  type: string
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error
  /getSites:
    get:
      summary: Retrieves the list of sites.
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error
  /getHomeLoadChart/{site_id}:
    get:
      summary: Retrieves the home load chart for a specific site.
      parameters:
        - in: path
          name: site_id
          required: true
          description: The ID of the site.
          schema:
            type: integer
        - in: query
          name: device_sn
          schema:
            type: string
          description: The device serial number (optional).
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error
  /getSiteDeviceParam/{site_id}:
    get:
      summary: Retrieves the device parameters for a specific site.
      parameters:
        - in: path
          name: site_id
          required: true
          description: The ID of the site.
          schema:
            type: integer
        - in: query
          name: paramType
          schema:
            type: string
          description: The type of parameter to retrieve. Default is "4".
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error
  /getRelateAndBindDevices:
    get:
      summary: Retrieves the related and bind devices for the user.
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error
  /getUserMqttInfo:
    get:
      summary: Retrieves the MQTT information for the user.
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error
  /getScenInfo/{site_id}:
    get:
      summary: Retrieves the scene information for a specific site.
      parameters:
        - in: path
          name: site_id
          required: true
          description: The ID of the site.
          schema:
            type: integer
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error
  /energyAnalysis/{site_id}:
    get:
      summary: Performs energy analysis for a specific site.
      parameters:
        - in: path
          name: site_id
          required: true
          description: The ID of the site.
          schema:
            type: integer
        - in: query
          name: device_sn
          schema:
            type: string
          description: The device serial number (optional).
        - in: query
          name: type
          schema:
            type: string
          description: The type of analysis to perform.
        - in: query
          name: start_time
          schema:
            type: string
          description: The start time for the analysis.
        - in: query
          name: end_time
          schema:
            type: string
          description: The end time for the analysis.
        - in: query
          name: device_type
          schema:
            type: string
          description: The type of device for the analysis. Default is "solar_production".
      responses:
        '200':
          description: Success response
        '500':
          description: Internal Server Error